Move C parts of `vec_ptype_finalise()` to `unspecified.c`
