+++ src/decl/ptype-decl.h
@@ -1,12 +1,6 @@
 static
 SEXP syms_vec_ptype;
 
-static
-SEXP syms_vec_ptype_finalise_dispatch;
-
-static
-SEXP fns_vec_ptype_finalise_dispatch;
-
 static inline
 r_obj* vec_ptype_slice(r_obj* x, r_obj* empty);
 
@@ -26,9 +20,3 @@ r_obj* vec_ptype_method(r_obj* x);
 
 static inline
 r_obj* vec_ptype_invoke(r_obj* x, r_obj* method);
-
-static
-r_obj* vec_ptype_finalise_unspecified(r_obj* x);
-
-static
-r_obj* vec_ptype_finalise_dispatch(r_obj* x);
