+++ src/vctrs.h
@@ -6,9 +6,6 @@
 
 // Vector types -------------------------------------------------
 
-SEXP vec_unspecified(R_len_t n);
-bool vec_is_unspecified(SEXP x);
-
 
 #include "type-info.h"
 
@@ -62,6 +59,7 @@ bool vec_is_unspecified(SEXP x);
 #include "typeof2-s3.h"
 #include "utils-dispatch.h"
 #include "utils.h"
+#include "unspecified.h"
 
 
 // Vector methods ------------------------------------------------
@@ -80,8 +78,6 @@ SEXP vec_proxy_unwrap(SEXP x);
 SEXP vec_slice_shaped(enum vctrs_type type, SEXP x, SEXP index);
 bool vec_requires_fallback(SEXP x, struct vctrs_proxy_info info);
 r_obj* vec_ptype(r_obj* x, struct vctrs_arg* x_arg, struct r_lazy call);
-SEXP vec_ptype_finalise(SEXP x);
-bool vec_is_unspecified(SEXP x);
 SEXP vec_names(SEXP x);
 SEXP vec_proxy_names(SEXP x);
 SEXP vec_group_loc(SEXP x);
